+++ CHANGELOG.md
@@ -4,6 +4,18 @@ If you like this project and find it useful, please consider giving it a star on
 
 All notable changes to this project will be documented in this file.
 
+## [1.2.3] - 2025-02-02
+
+### Changed
+
+- [package]: Requires matterbridge 2.1.0.
+- [package]: Updated package.
+- [package]: Updated dependencies.
+
+<a href="https://www.buymeacoffee.com/luligugithub">
+  <img src="./yellow-button.png" alt="Buy me a coffee" width="120">
+</a>
+
 ## [1.2.2] - 2024-12-23
 
 ### Added